second util function version for double

+++ src/crystalizer_ui.cpp
@@ -168,13 +168,13 @@ void CrystalizerUi::build_bands(const int& nbands) {
 }
 
 void CrystalizerUi::on_new_range_before(double value) {
-  range_before->set_value(util::db_to_linear(static_cast<float>(value)));
+  range_before->set_value(util::db_to_linear(value));
 
   range_before_label->set_text(level_to_str(value, 2));
 }
 
 void CrystalizerUi::on_new_range_after(double value) {
-  range_after->set_value(util::db_to_linear(static_cast<float>(value)));
+  range_after->set_value(util::db_to_linear(value));
 
   range_after_label->set_text(level_to_str(value, 2));
 }
